docs: Document `PerformanceTrigger` approach (#596)
