+++ modyn/supervisor/internal/triggers/datadrifttrigger.py
@@ -5,15 +5,15 @@ import logging
 from collections.abc import Generator
 
 from modyn.config.schema.pipeline import DataDriftTriggerConfig
+from modyn.config.schema.pipeline.trigger.drift.criterion import (
+    DynamicPercentileThresholdCriterion,
+    DynamicRollingAverageThresholdCriterion,
+    ThresholdDecisionCriterion,
+)
 from modyn.config.schema.pipeline.trigger.drift.detection_window import (
     AmountWindowingStrategy,
     DriftWindowingStrategy,
     TimeWindowingStrategy,
-)
-from modyn.config.schema.pipeline.trigger.drift.metric import (
-    DynamicPercentileThresholdCriterion,
-    DynamicRollingAverageThresholdCriterion,
-    ThresholdDecisionCriterion,
 )
 from modyn.config.schema.pipeline.trigger.drift.result import MetricResult
 from modyn.supervisor.internal.triggers.drift.decision_policy import (
